Extract offerings code to different class while keeping tests
