+++ purchases/src/main/kotlin/com/revenuecat/purchases/Purchases.kt
@@ -23,7 +23,6 @@ import com.revenuecat.purchases.common.BillingAbstract
 import com.revenuecat.purchases.common.Config
 import com.revenuecat.purchases.common.Dispatcher
 import com.revenuecat.purchases.common.LogIntent
-import com.revenuecat.purchases.common.OfferingParser
 import com.revenuecat.purchases.common.PlatformInfo
 import com.revenuecat.purchases.common.ReceiptInfo
 import com.revenuecat.purchases.common.ReplaceProductInfo
@@ -34,6 +33,7 @@ import com.revenuecat.purchases.common.diagnostics.DiagnosticsSynchronizer
 import com.revenuecat.purchases.common.errorLog
 import com.revenuecat.purchases.common.infoLog
 import com.revenuecat.purchases.common.log
+import com.revenuecat.purchases.common.offerings.OfferingsManager
 import com.revenuecat.purchases.common.offlineentitlements.OfflineEntitlementsManager
 import com.revenuecat.purchases.common.sha1
 import com.revenuecat.purchases.common.subscriberattributes.SubscriberAttributeKey
@@ -65,8 +65,6 @@ import com.revenuecat.purchases.strings.OfferingStrings
 import com.revenuecat.purchases.strings.PurchaseStrings
 import com.revenuecat.purchases.strings.RestoreStrings
 import com.revenuecat.purchases.subscriberattributes.SubscriberAttributesManager
-import org.json.JSONException
-import org.json.JSONObject
 import java.net.URL
 import java.util.Collections.emptyMap
 
@@ -93,12 +91,12 @@ class Purchases internal constructor(
     private val subscriberAttributesManager: SubscriberAttributesManager,
     @set:JvmSynthetic @get:JvmSynthetic internal var appConfig: AppConfig,
     private val customerInfoHelper: CustomerInfoHelper,
-    private val offeringParser: OfferingParser,
     diagnosticsSynchronizer: DiagnosticsSynchronizer?,
     @get:VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
     internal val offlineEntitlementsManager: OfflineEntitlementsManager,
     private val postReceiptHelper: PostReceiptHelper,
     private val syncPurchasesHelper: SyncPurchasesHelper,
+    private val offeringsManager: OfferingsManager,
     // This is nullable due to: https://github.com/RevenueCat/purchases-flutter/issues/408
     private val mainHandler: Handler? = Handler(Looper.getMainLooper())
 ) : LifecycleDelegate {
@@ -207,7 +205,7 @@ class Purchases internal constructor(
         }
         if (deviceCache.isOfferingsCacheStale(appInBackground = false)) {
             log(LogIntent.DEBUG, OfferingStrings.OFFERINGS_STALE_UPDATING_IN_FOREGROUND)
-            fetchAndCacheOfferings(identityManager.currentAppUserID, appInBackground = false)
+            offeringsManager.fetchAndCacheOfferings(identityManager.currentAppUserID, appInBackground = false)
             log(LogIntent.RC_SUCCESS, OfferingStrings.OFFERINGS_UPDATED_FROM_NETWORK)
         }
         offlineEntitlementsManager.updateProductEntitlementMappingCacheIfStale()
@@ -327,7 +325,12 @@ class Purchases internal constructor(
         }
         if (cachedOfferings == null) {
             log(LogIntent.DEBUG, OfferingStrings.NO_CACHED_OFFERINGS_FETCHING_NETWORK)
-            fetchAndCacheOfferings(appUserID, state.appInBackground, listener)
+            offeringsManager.fetchAndCacheOfferings(
+                appUserID,
+                state.appInBackground,
+                { listener.onReceived(it) },
+                { listener.onError(it) }
+            )
         } else {
             log(LogIntent.DEBUG, OfferingStrings.VENDING_OFFERINGS_CACHE)
             dispatch {
@@ -340,7 +343,7 @@ class Purchases internal constructor(
                         if (appInBackground) OfferingStrings.OFFERINGS_STALE_UPDATING_IN_BACKGROUND
                         else OfferingStrings.OFFERINGS_STALE_UPDATING_IN_FOREGROUND
                     )
-                    fetchAndCacheOfferings(appUserID, appInBackground)
+                    offeringsManager.fetchAndCacheOfferings(appUserID, appInBackground)
                     log(LogIntent.RC_SUCCESS, OfferingStrings.OFFERINGS_UPDATED_FROM_NETWORK)
                 }
             }
@@ -638,7 +641,7 @@ class Purchases internal constructor(
                         callback?.onReceived(customerInfo, created)
                         customerInfoHelper.sendUpdatedCustomerInfoToDelegateIfChanged(customerInfo)
                     }
-                    fetchAndCacheOfferings(newAppUserID, state.appInBackground)
+                    offeringsManager.fetchAndCacheOfferings(newAppUserID, state.appInBackground)
                 },
                 onError = { error ->
                     dispatch { callback?.onError(error) }
@@ -1073,116 +1076,6 @@ class Purchases internal constructor(
 
     // region Private Methods
 
-    private fun fetchAndCacheOfferings(
-        appUserID: String,
-        appInBackground: Boolean,
-        completion: ReceiveOfferingsCallback? = null
-    ) {
-        deviceCache.setOfferingsCacheTimestampToNow()
-        backend.getOfferings(
-            appUserID,
-            appInBackground,
-            { offeringsJSON ->
-                try {
-                    val allRequestedProductIdentifiers = extractProductIdentifiers(offeringsJSON)
-                    if (allRequestedProductIdentifiers.isEmpty()) {
-                        handleErrorFetchingOfferings(
-                            PurchasesError(
-                                PurchasesErrorCode.ConfigurationError,
-                                OfferingStrings.CONFIGURATION_ERROR_NO_PRODUCTS_FOR_OFFERINGS
-                            ),
-                            completion
-                        )
-                    } else {
-                        getStoreProductsById(allRequestedProductIdentifiers, { productsById ->
-                            logMissingProducts(allRequestedProductIdentifiers, productsById)
-
-                            val offerings = offeringParser.createOfferings(offeringsJSON, productsById)
-                            if (offerings.all.isEmpty()) {
-                                handleErrorFetchingOfferings(
-                                    PurchasesError(
-                                        PurchasesErrorCode.ConfigurationError,
-                                        OfferingStrings.CONFIGURATION_ERROR_PRODUCTS_NOT_FOUND
-                                    ),
-                                    completion
-                                )
-                            } else {
-                                synchronized(this@Purchases) {
-                                    deviceCache.cacheOfferings(offerings)
-                                }
-                                dispatch {
-                                    completion?.onReceived(offerings)
-                                }
-                            }
-                        }, { error ->
-                            handleErrorFetchingOfferings(error, completion)
-                        })
-                    }
-                } catch (error: JSONException) {
-                    log(LogIntent.RC_ERROR, OfferingStrings.JSON_EXCEPTION_ERROR.format(error.localizedMessage))
-                    handleErrorFetchingOfferings(
-                        PurchasesError(
-                            PurchasesErrorCode.UnexpectedBackendResponseError,
-                            error.localizedMessage
-                        ),
-                        completion
-                    )
-                }
-            }, { error ->
-                handleErrorFetchingOfferings(error, completion)
-            })
-    }
-
-    private fun extractProductIdentifiers(offeringsJSON: JSONObject): Set<String> {
-        val jsonOfferingsArray = offeringsJSON.getJSONArray("offerings")
-        val productIds = mutableSetOf<String>()
-        for (i in 0 until jsonOfferingsArray.length()) {
-            val jsonPackagesArray =
-                jsonOfferingsArray.getJSONObject(i).getJSONArray("packages")
-            for (j in 0 until jsonPackagesArray.length()) {
-                jsonPackagesArray.getJSONObject(j)
-                    .optString("platform_product_identifier").takeIf { it.isNotBlank() }?.let {
-                        productIds.add(it)
-                    }
-            }
-        }
-        return productIds
-    }
-
-    private fun handleErrorFetchingOfferings(
-        error: PurchasesError,
-        completion: ReceiveOfferingsCallback?
-    ) {
-        val errorCausedByPurchases = setOf(
-            PurchasesErrorCode.ConfigurationError,
-            PurchasesErrorCode.UnexpectedBackendResponseError
-        )
-            .contains(error.code)
-
-        log(
-            if (errorCausedByPurchases) LogIntent.RC_ERROR else LogIntent.GOOGLE_ERROR,
-            OfferingStrings.FETCHING_OFFERINGS_ERROR.format(error)
-        )
-
-        deviceCache.clearOfferingsCacheTimestamp()
-        dispatch {
-            completion?.onError(error)
-        }
-    }
-
-    private fun logMissingProducts(
-        allProductIdsInOfferings: Set<String>,
-        storeProductByID: Map<String, List<StoreProduct>>
-    ) = allProductIdsInOfferings
-        .filterNot { storeProductByID.containsKey(it) }
-        .takeIf { it.isNotEmpty() }
-        ?.let { missingProducts ->
-            log(
-                LogIntent.GOOGLE_WARNING, OfferingStrings.CANNOT_FIND_PRODUCT_CONFIGURATION_ERROR
-                    .format(missingProducts.joinToString(", "))
-            )
-        }
-
     private fun getProductsOfTypes(
         productIds: Set<String>,
         types: Set<ProductType>,
@@ -1235,7 +1128,7 @@ class Purchases internal constructor(
                 appInBackground,
                 completion
             )
-            fetchAndCacheOfferings(appUserID, appInBackground)
+            offeringsManager.fetchAndCacheOfferings(appUserID, appInBackground)
         }
     }
 
@@ -1292,40 +1185,6 @@ class Purchases internal constructor(
                 )
             }
         }
-    }
-
-    private fun getStoreProductsById(
-        productIds: Set<String>,
-        onCompleted: (Map<String, List<StoreProduct>>) -> Unit,
-        onError: (PurchasesError) -> Unit
-    ) {
-        billing.queryProductDetailsAsync(
-            ProductType.SUBS,
-            productIds,
-            { subscriptionProducts ->
-                val productsById = subscriptionProducts
-                    .groupBy { subProduct -> subProduct.purchasingData.productId }
-                    .toMutableMap()
-                val subscriptionIds = productsById.keys
-
-                val inAppProductIds = productIds - subscriptionIds
-                if (inAppProductIds.isNotEmpty()) {
-                    billing.queryProductDetailsAsync(
-                        ProductType.INAPP,
-                        inAppProductIds,
-                        { inAppProducts ->
-                            productsById.putAll(inAppProducts.map { it.purchasingData.productId to listOf(it) })
-                            onCompleted(productsById)
-                        }, {
-                            onError(it)
-                        }
-                    )
-                } else {
-                    onCompleted(productsById)
-                }
-            }, {
-                onError(it)
-            })
     }
 
     private fun dispatch(action: () -> Unit) {
